allow node 6
